+++ changelog.md
@@ -30,6 +30,7 @@
 - Added `data-invariant` attribute to number input controls 
 - Closed #543 Google Category (GMC) always get lost when copying a product
 - Ajax request to external URLs do not add X-XSRF-Tokens anymore
+- Definition of several preselected product variant attribute values is now prohibited
 
 ### Bugfixes